Fixes in SAM processing

+++ source/libs/fio/sam-record.cc
@@ -1,5 +1,6 @@
 #include "sam-record.h"
 
+#include <iostream>
 #include <sstream>
 
 namespace fio {
@@ -8,12 +9,12 @@ namespace fio {
             : qname(fields[0]),
               flag(static_cast<uint16_t>(std::stoi(fields[1]))),
               rname(fields[2]),
-              pos((uint32_t) std::stoi(fields[3])),
-              mapq((uint8_t) std::stoi(fields[4])),
+              pos(static_cast<uint32_t>(std::stoi(fields[3]))),
+              mapq(static_cast<uint8_t>(std::stoi(fields[4]))),
               cigar(fields[5]),
               rnext(fields[6]),
-              pnext((uint32_t) std::stoi(fields[7])),
-              tlen((int64_t) std::stoi(fields[8])),
+              pnext(static_cast<uint32_t>(std::stoi(fields[7]))),
+              tlen(static_cast<int64_t>(std::stoi(fields[8]))),
               seq(fields[9]),
               qual(fields[10]),
               opt(fields[11]) {}
@@ -27,7 +28,7 @@ namespace fio {
         ss << flag << " ";
         ss << rname << " ";
         ss << pos << " ";
-        ss << mapq << " ";
+        ss << static_cast<int>(mapq) << " ";  // cast needed because std::cout is overloaded for uint8_t
         ss << cigar << " ";
         ss << rnext << " ";
         ss << pnext << " ";
